Remove the content API toggle
