+++ catalogue/webapp/pages/search/stories.tsx
@@ -16,12 +16,10 @@ import { NextPageWithLayout } from '@weco/common/views/pages/_app';
 import { removeUndefinedProps } from '@weco/common/utils/json';
 import { AppErrorProps } from '@weco/common/services/app';
 import { getServerData } from '@weco/common/server-data';
-import { getStories } from '@weco/catalogue/services/prismic/fetch/articles';
 import { Pageview } from '@weco/common/services/conversion/track';
 import { pluralize } from '@weco/common/utils/grammar';
 import { getQueryPropertyValue } from '@weco/common/utils/search';
 import { getArticles } from '@weco/catalogue/services/wellcome/content/articles';
-import { useToggles } from '@weco/common/server-data/Context';
 
 // Types
 import {
@@ -60,26 +58,20 @@ export const SearchPage: NextPageWithLayout<Props> = ({
   query,
 }) => {
   const { query: queryString } = query;
-  const { contentApi } = useToggles();
   const returnedStories = storyResponseList || newStoryResponseList;
 
   const sortOptions = [
     // Default value to be left empty as to not be reflected in URL query
-    ...(contentApi
-      ? [
-          {
-            value: '',
-            text: 'Relevance',
-          },
-        ]
-      : []),
-
+    {
+      value: '',
+      text: 'Relevance',
+    },
     {
       value: 'publication.dates.asc',
       text: 'Oldest to newest',
     },
     {
-      value: contentApi ? 'publication.dates.desc' : '',
+      value: 'publication.dates.desc',
       text: 'Newest to oldest',
     },
   ];
@@ -103,16 +95,12 @@ export const SearchPage: NextPageWithLayout<Props> = ({
                     options={sortOptions}
                     jsLessOptions={{
                       sort: [
-                        ...(contentApi
-                          ? [
-                              {
-                                value: '',
-                                text: 'Relevance',
-                              },
-                            ]
-                          : []),
                         {
-                          value: contentApi ? 'publication.dates' : '',
+                          value: '',
+                          text: 'Relevance',
+                        },
+                        {
+                          value: 'publication.dates',
                           text: 'Publication dates',
                         },
                       ],
@@ -182,7 +170,6 @@ export const getServerSideProps: GetServerSideProps<
 > = async context => {
   const serverData = await getServerData(context);
   const query = context.query;
-  const { contentApi } = serverData.toggles;
   const defaultProps = removeUndefinedProps({
     serverData,
     storyResponseList: { totalResults: 0 },
@@ -198,38 +185,18 @@ export const getServerSideProps: GetServerSideProps<
   const { page, ...restOfQuery } = query;
   const pageNumber = page !== '1' && getQueryPropertyValue(page);
 
-  // Setting a default order of descending publication date as default state
-  // as the Prismic default is by "last updated"
+  const storyResponseList = await getArticles({
+    params: {
+      ...restOfQuery,
+      sort: getQueryPropertyValue(query.sort),
+      sortOrder: getQueryPropertyValue(query.sortOrder),
+      ...(pageNumber && { page: Number(pageNumber) }),
+    },
+    pageSize: 6,
+    toggles: serverData.toggles,
+  });
 
-  const storyResponseList = contentApi
-    ? undefined
-    : await getStories({
-        query: {
-          ...restOfQuery,
-          sort: getQueryPropertyValue(query.sort) || 'publication.dates',
-          sortOrder: getQueryPropertyValue(query.sortOrder) || 'desc',
-          ...(pageNumber && { page: pageNumber }),
-        },
-        pageSize: 6,
-      });
-
-  const newStoryResponseList = contentApi
-    ? await getArticles({
-        params: {
-          ...restOfQuery,
-          sort: getQueryPropertyValue(query.sort),
-          sortOrder: getQueryPropertyValue(query.sortOrder),
-          ...(pageNumber && { page: Number(pageNumber) }),
-        },
-        pageSize: 6,
-        toggles: serverData.toggles,
-      })
-    : undefined;
-
-  if (
-    storyResponseList?.type === 'Error' ||
-    newStoryResponseList?.type === 'Error'
-  ) {
+  if (storyResponseList?.type === 'Error') {
     // Prismic returns Internal Server Errors without much details for certain queries, such as query=t:"PP.PRE.D.1.1"
     // As this is a temporary search tool until we replace it with our own API, we'll just return No Result should it happen
     return {
@@ -256,28 +223,17 @@ export const getServerSideProps: GetServerSideProps<
     props: removeUndefinedProps({
       ...defaultProps,
       storyResponseList,
-      newStoryResponseList,
       pageview: {
         name: 'stories',
         properties: {
-          totalResults: newStoryResponseList
-            ? newStoryResponseList?.type === 'ResultList'
-              ? newStoryResponseList.results
-              : 0
-            : storyResponseList?.type === 'ResultList'
-            ? storyResponseList.totalResults
-            : 0,
+          totalResults: storyResponseList?.results ?? 0,
         },
       },
-      apiToolbarLinks: newStoryResponseList
-        ? [
-            {
-              id: 'content-api',
-              label: 'Content API query',
-              link: newStoryResponseList._requestUrl,
-            },
-          ]
-        : [],
+      apiToolbarLinks: {
+        id: 'content-api',
+        label: 'Content API query',
+        link: storyResponseList._requestUrl,
+      },
     }),
   };
 };
